decode: fix bit_read_UMC
@arturred at GH #386 was right, UMC may need up to 8 byte.
The #662 testcase works now, and the samples also.
Fixes GH #662 and #126. ERR-2 from #386 not yet.

+++ src/decode_r11.c
@@ -211,7 +211,7 @@ decode_preR13_section_hdr (const char *restrict name, Dwg_Section_Type_r11 id,
   if (tbl->number > 0 && (tbl->address + (tbl->number * tbl->size) > dat->size))
     {
       LOG_ERROR ("%s.size overflow %lu > %lu", tbl->name,
-                 tbl->address + (tbl->number * tbl->size), dat->size);
+                 (unsigned long)(tbl->address + (tbl->number * tbl->size)), (unsigned long)dat->size);
       return DWG_ERR_INVALIDDWG;
     }
   return 0;
